refactor: MenuProducts 1급 컬렉션으로 변경.
