+++ src/main/java/kitchenpos/menu/domain/MenuProducts.java
@@ -1,0 +1,35 @@
+package kitchenpos.menu.domain;
+
+import java.math.BigDecimal;
+import java.util.List;
+import javax.persistence.CascadeType;
+import javax.persistence.Embeddable;
+import javax.persistence.OneToMany;
+
+@Embeddable
+public class MenuProducts {
+
+    @OneToMany(mappedBy = "menu", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<MenuProduct> menuProducts;
+
+    public MenuProducts() {
+
+    }
+
+    public MenuProducts(Menu menu, List<MenuProduct> menuProducts) {
+        this.menuProducts = menuProducts;
+        menuProducts.forEach(menuProduct -> menuProduct.setMenu(menu));
+    }
+
+    public BigDecimal sumOfProducts() {
+        BigDecimal sum = BigDecimal.ZERO;
+        for (final MenuProduct menuProduct : menuProducts) {
+            sum = sum.add(menuProduct.getTotalPrice());
+        }
+        return sum;
+    }
+
+    public List<MenuProduct> getMenuProducts() {
+        return menuProducts;
+    }
+}
